fix grading bugs and update ui
