+++ src/src/app.d.ts
@@ -25,6 +25,32 @@ declare global {
 		content: string;
 		createdBy: string;
 	}
+
+	export interface GradeData {
+		_id: string;
+		studentID: string;
+		quizID: string;
+		workspaceID: string;
+		mark: number;
+	}
+
+	export interface QuizData {
+		_id: string;
+		title: string;
+		totalPossibleScore: number;
+	}
+
+	export interface AssessmentStat {
+		name: string;
+		submitted: number;
+		average: number;
+		passRate: number;
+	}
+
+	export interface StudentGrade {
+		name: string;
+		grades: number[];
+	}
 }
 
 export {};
